refactor: unify prompt envelope + ui metadata helpers
