+++ src/lib/text/message-metadata.ts
@@ -1,3 +1,26 @@
+export type AgentInstructionParams = {
+  workspacePath: string;
+  message: string;
+};
+
+// Keep this phrase as a stable sentinel that both the envelope builder and
+// metadata stripper can agree on. If we ever change this string, we should
+// update the strip regexes below (or better: replace them with a structured
+// format).
+const ENVELOPE_SENTINEL = "memory_search.";
+
+export const buildAgentInstruction = ({
+  workspacePath,
+  message,
+}: AgentInstructionParams): string => {
+  const trimmed = message.trim();
+  if (!trimmed) return trimmed;
+  if (trimmed.startsWith("/")) return trimmed;
+  const workspace = workspacePath?.trim();
+  if (!workspace) return trimmed;
+  return `Workspace path: ${workspace}. Operate within this repository. You may also read/write your agent workspace files (AGENTS.md, SOUL.md, IDENTITY.md, USER.md, HEARTBEAT.md, TOOLS.md, MEMORY.md). Use MEMORY.md or memory/*.md directly for durable memory; do not rely on ${ENVELOPE_SENTINEL}\n\n${trimmed}`;
+};
+
 const PROJECT_PROMPT_BLOCK_RE = /^(?:Project|Workspace) path:[\s\S]*?\n\s*\n/i;
 const PROJECT_PROMPT_INLINE_RE = /^(?:Project|Workspace) path:[\s\S]*?memory_search\.\s*/i;
 const RESET_PROMPT_RE =
